separation of main (lession 27.10.2024)
