+++ Mirror/CMakeLists.txt
@@ -1,3 +1,4 @@
 project( File_work LANGUAGES CXX )
 
+set(CMAKE_BUILD_TYPE Debug)
 add_executable( Mirror Mirror.cpp Checker.cpp MakeTask.cpp)
