+++ Mirror/MakeTask.h
@@ -4,6 +4,13 @@
 
 #include <memory>
 
+class Task
+{
+public:
+	virtual ~Task() = default;
+	virtual void Exec() = 0;
+};
+
 std::unique_ptr<Task> MakeTask(int argc, char** argv);
 
 #endif
